Hide NavigationBar
If the mainViewController is embedded in UINavigationController, then the PMParallax View will look weird because of the top bar. So, we need to set the navigationBar to hidden in viewDidLoad, and then enable it after closing the PMParallax View.
User should be careful with the Auto-layout of buttons and views, which should be aligned according to Navigation bars.